+++ tests/testthat/test-DataSurvival.R
@@ -105,6 +105,24 @@ test_that("DataSurvival print method works as expected", {
     })
 })
 
+test_that("model.matrix for DataSurvival works as expected", {
+    set.seed(123)
+    x <- data.frame(
+        vpt = c("b", "a", "c", "d", "e"),
+        vtime = c(10, 20, 30, 25, 15),
+        vevent = c(1, 1, 0, 1, 0),
+        vcov1 = c("A", "A", "B", "B", "A"),
+        vcov2 = rnorm(5)
+    )
+
+    obj <- DataSurvival(
+        data = x,
+        formula = Surv(vtime, vevent) ~ vcov1 * vcov2
+    )
+
+    res <- model.matrix(obj)
+    expect_snapshot_value(res, tolerance = 1e-2, style = "deparse")
+})
 
 test_that("mirror_design_matrix() works as expected", {
     set.seed(3102)