+++ tests/testthat/test-Prior.R
@@ -5,7 +5,7 @@ test_that("Priors work as expected", {
             initialValues(x),
             4 * 0.5
         ),
-        local_rnorm = \(...) 0
+        local_rnorm = \(n, ...) rep(0, n)
     )
     expect_equal(
         as.StanModule(x, name = "bob"),
@@ -22,7 +22,7 @@ test_that("Priors work as expected", {
             initialValues(x),
             exp(log(4) + 2) * 0.5
         ),
-        local_rlnorm = \(...) 0
+        local_rlnorm = \(n, ...) rep(0, n)
     )
     expect_equal(
         as.StanModule(x, name = "tim"),
@@ -181,7 +181,7 @@ test_that("jmpost.prior_shrinkage works as expected", {
             ## Reset Shrinkage factor
             options("jmpost.prior_shrinkage" = 0.5)
         },
-        local_rnorm = \(...) 4
+        local_rnorm = \(n, ...) rep(4, n)
     )
 })
 
@@ -268,5 +268,33 @@ test_that("Parameters in priors must be length 1 #422", {
     expect_error(
         prior_gamma(c(1, 2), 2),
         "Parameter `alpha`"
+    )
+})
+
+test_that("prior_normal_vector works as expected", {
+    x <- prior_normal_vector(c(1, 2, 3), c(4, 5, 6))
+    x_char <- as.character(x)
+
+    expect_equal(
+        x_char,
+        "normal(mus = [1, 2, 3], sigmas = [4, 5, 6])"
+    )
+
+    x_inits <- initialValues(x)
+    expect_numeric(x_inits, len = 3)
+
+    x_stan_module <- as.StanModule(x, name = "bob")
+    expect_equal(
+        x_stan_module,
+        StanModule(test_path("models", "Prior_4.stan"))
+    )
+    expect_equal(
+        as_stan_list(x, name = "bob"),
+        list(
+            prior_mus_bob = c(1, 2, 3),
+            prior_sigmas_bob = c(4, 5, 6),
+            prior_dim_mus_bob = 3,
+            prior_dim_sigmas_bob = 3
+        )
     )
 })