+++ tests/testthat/test-misc_models.R
@@ -3,7 +3,7 @@ test_data_1 <- ensure_test_data_1()
 test_that("Longitudinal Model doesn't print sampler rejection messages", {
     # These rejections typically happen when the sampler samples a
     # 0 value for the variance parameter. Sensible initial values +
-    # setting near 0 limits (as opposed to 0) should avoid this
+    # setting near 0 limits (as opposed to 0) should avoid this.
 
     mp <- capture_messages({
         devnull_out <- capture.output({
@@ -14,7 +14,8 @@ test_that("Longitudinal Model doesn't print sampler rejection messages", {
                 iter_warmup = 3,
                 chains = 1,
                 refresh = 0,
-                parallel_chains = 1
+                parallel_chains = 1,
+                seed = 324
             )
         })
     })