+++ NEWS.md
@@ -1,4 +1,3 @@
-
 # jmpost (development version)
 
 - `LongitudinalModel` objects now have a new slot `scaled_variance` which stores the used variance option, and this is part of the `print` output now thereby transparently communicating to the user the choice of the multiplicative or additive error model.
@@ -7,11 +6,13 @@
 - Included new `LongitudinalRandomEffects()` function which can be used to extract the patient-level random effects parameter samples from a `JointModelSample` object (#423).
 - Introduced the `saveObject()` method for `JointModelSample` objects in order to serialise them to disk (#431).
 - Added support for truncated prior distributions e.g. you can now apply a normal prior to a strictly positive parameter and jmpost will take care of adjusting the density accordingly (#429).
+- Added `prior_normal_vector()` which can be used to assign different normal prior distributions to the coefficients of the survival model covariates.
 - Included new Gamma distribution survival model (#411).
 - Reworked LOO calculations to apply to each individual submodel and disabled LOO calculations for the overall joint model (#402).
 - Added support for additive variance (#403).
 - Added support for independent variances per study/arm (#389).
 - Miscellaneous bug fixes.
+- Introduce new package options for bounds close to zero which are used to avoid MCMC sampler starting warnings.
 
 # jmpost 0.0.1